requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Redis
redis==5.0.1
aioredis==2.0.1

# HTTP client for service communication
httpx==0.25.2

# Validation and utilities
email-validator==2.1.0
python-phonenumbers==8.13.26

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Environment
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Health checks
psutil==5.9.6